Don't include json_loader.h in ir.h
